static-pods: migrate static-pods tool to use configuration file and remove conditional compilation
